Refactored request component and updated validation on signup page
