+++ src/views/manage/AuctionDetail.html
@@ -58,36 +58,13 @@
                         Create date: YYYY/MM/DD hh:mm:ss
                     </div>
                     <div class="auction-detail-content-row4">
-                        <label for="create-by-input" class="create-by-label">Create By</label>
-                        <input type="text" class="create-by-input" value="Thinh">
+                        <label for="start-date-input" class="start-date-label">Start Date</label>
+                        <input type="datetime" class="start-date-input" value=" YYYY/MM/DD hh:mm:ss">
                     </div>
                     <div class="auction-detail-content-row5">
-                        <div class="update-by">
-                            <label for="update-by-input" class="update-by-label">Update By</label>
-                            <input type="text" class="update-by-input" value="Thinh">
-                        </div>
-                        <div class="last-update"> <label for="last-update-input" class="last-update-label">Last
-                                Update</label>
-                            <input type="datetime" class="last-update-input" value="YYYY/MM/DD hh:mm:ss">
-                        </div>
+                        <div class="add-fish"><i class="fa-solid fa-plus"></i></div>
                     </div>
                     <div class="auction-detail-content-row6">
-                        <div class="start-date">
-                            <label for="start-date-input" class="start-date-label">Start Date</label>
-                            <input type="datetime" class="start-date-input" value="YYYY/MM/DD hh:mm:ss">
-                        </div>
-                        <div class="end-date">
-                            <label for="end-date-input" class="end-date-label">End Date</label>
-                            <input type="datetime" class="end-date-input" value="YYYY/MM/DD hh:mm:ss">
-                        </div>
-                    </div>
-                    <div class="auction-detail-content-row7">
-                        <div class="fish-entry">
-                            <div class="fish-entry-text">Fish Entry</div>
-                            <button class="fish-entry-addition"><i class="fa-solid fa-plus"></i></button>
-                        </div>
-                    </div>
-                    <div class="auction-detail-content-row8">
                         <table class="table-manage-blog">
                             <tr>
                                 <th>No</th>
@@ -113,7 +90,7 @@
 
 
 
-                    <div class="auction-detail-content-row10">
+                    <div class="auction-detail-content-row7">
                         <button class="update-btn">Update</button>
                         <button class="cancel-btn">Cancel</button>
                     </div>
