+++ src/views/manage/MemberDetail.html
@@ -1,0 +1,153 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>MemberDetail</title>
+    <link rel="stylesheet" href="MemberDetail.css">
+    <link
+        href="https://fonts.googleapis.com/css2?family=Saira+Condensed:wght@100;200;300;400;500;600;700;800;900&display=swap"
+        rel="stylesheet">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.6.0/css/all.min.css"
+        integrity="sha512-Kc323vGBEqzTmouAECnVceyQqyqdsSiqLQISBL29aUW4U/M7pSPA/gEUZQqv1cwx4OnYxTxve5UMg5GT6L4JJg=="
+        crossorigin="anonymous" referrerpolicy="no-referrer" />
+</head>
+
+<body>
+    <div class="member-detail-container">
+        <div class="header">
+            <div class="navigation-bar">
+                <div class="navigation-bar-left-content">
+                    <img class="logo" src="../../assets/images/logo_png.png" alt="">
+                    <div class="project-name">Nishikigoi Nexus</div>
+                </div>
+
+                <div class="navigation-bar-right-content">
+                    <a class="staff">STAFF</a>
+                    <a class="home">HOME</a>
+                    <a class="auction">AUCTION</a>
+                    <a class="blog">BLOG</a>
+                    <a class="about">ABOUT</a>
+                    <a class="account">ACCOUNT</a>
+                </div>
+            </div>
+        </div>
+        <div class="body-content">
+            <div class="navigation-bar-vertically">
+                <a class="member">Member</a>
+                <a class="breeder">Breeder</a>
+                <a class="request">Request</a>
+                <a class="auction-vertically">Auction</a>
+                <a class="koi">KOI</a>
+                <a class="blog-vertically">Blog</a>
+            </div>
+            <div class="body-content-right">
+                <div class="member-detail-content">
+                    <div class="member-detail-content-row1">
+                        <div class="status">Status: Active</div>
+                        <select name="" id="" class="set-status">
+                            <option value="">Set status &nbsp;&nbsp; v</option>
+
+                        </select>
+                    </div>
+                    <div class="member-detail-content-row2">
+                        Member Detail
+                    </div>
+                    <div class="member-detail-content-row3">
+                        Create date: YYYY/MM/DD hh:mm:ss
+                    </div>
+
+                    <div class="member-detail-content-row4">
+                        <div class="email">
+                            <label for="email-input" class="email-label">Email</label>
+                            <input type="text" class="email-input" value="thinhdlse181755@fpt.edu.vn">
+                        </div>
+                        <div class="gender"> <label for="gender-input" class="gender-label">Gender</label>
+                            <input type="text" class="gender-input" value="0362xxx366">
+                        </div>
+                    </div>
+                    <div class="member-detail-content-row5">
+                        <div class="first-name">
+                            <label for="first-name-input" class="first-name-label">First Name</label>
+                            <input type="text" class="first-name-input" value="Thinh">
+                        </div>
+                        <div class="last-name"> <label for="last-name-input" class="last-name-label">Last Name</label>
+                            <input type="text" class="last-name-input" value="Dinh">
+                        </div>
+                    </div>
+                    <div class="member-detail-content-row6">
+                        <div class="phone"> <label for="phone-input" class="phone-label">Phone</label>
+                            <input type="text" class="phone-input" value="0362xxx366">
+                        </div>
+                        <div class="city">
+                            <label for="city-input" class="city-label">City</label>
+                            <input type="text" class="city-input" value="Tra Vinh">
+                        </div>
+                    </div>
+                    <div class="member-detail-content-row7">
+                        <label for="address-input" class="address-label">Address</label>
+                        <input type="text" class="address-input" value="Tra Vinh">
+                    </div>
+                    <div class="member-detail-content-row8">
+                        <div class="ban-unban">Ban / Unban</div>
+                    </div>
+                    <div class="member-detail-content-row9">
+                        <label for="reason-input" class="reason-label">Reason</label>
+                        <input type="text" class="reason-input" value="Customer didn't pay">
+                    </div>
+                    <div class="member-detail-content-row10">
+                        <button class="ban-btn">Ban</button>
+                        <button class="unban-btn">Unban</button>
+                    </div>
+
+                    <div class="member-detail-content-row11">
+                        <div class="bidding-history">Bidding History</div>
+                    </div>
+                    <div class="member-detail-content-row12">
+                        <table class="table-request-history">
+                            <tr>
+                                <th>No.</th>
+                                <th>Fish Id</th>
+                                <th>Bidding date</th>
+                                <th>Amount bidding</th>
+
+
+
+                            </tr>
+                            <tr>
+                                <td>1</td>
+                                <td>fish 1</td>
+                                <td>YYYY/MM/DD hh:mm:ss</td>
+                                <td>1234</td>
+                            </tr>
+                            <tr>
+                                <td>1</td>
+                                <td>fish 1</td>
+                                <td>YYYY/MM/DD hh:mm:ss</td>
+                                <td>1234</td>
+                            </tr>
+                            <tr>
+                                <td>1</td>
+                                <td>fish 1</td>
+                                <td>YYYY/MM/DD hh:mm:ss</td>
+                                <td>1234</td>
+                            </tr>
+                        </table>
+                    </div>
+                </div>
+
+            </div>
+
+
+        </div>
+
+    </div>
+
+
+
+
+
+</body>
+
+</html>
